QBluetoothUuid: inherit all ctors from QUuid
QBluetoothUuid clearly is-a QUuid, so it should have the same
constructors, though it may add its own on top. Three Qt releases
(6.3-6.6) of QUuid changes have, however, not been properly reflected
in QBluetoothUuid, so fix it once and for all by adding

    using QUuid::QUuid;

to QBluetoothUuid, thereby inheriting all the QUuid ctors (except
SMFs) that QBluethoothUuid doesn't supply itself.

This fixes or simplifies:

- The Id128Bytes ctor, which we don't need to supply at all now.
- The QString ctor, which was replaced by a QAnyStringView overload in
  QUuid a long time ago (6.3).
- Anything else that we have missed in the past (GUID, e.g.).

The superfluous constructors still have to be retained in the ABI,
because QBluetoothUuid is unfortunately exported wholesale. This
affects only the QString ctor. The other one was not in 6.5, and we
don't provide BC with pre-releases of Qt.

Add some variance to the fromString() tests now that we have
QAnyStringView instead of just QString.

Found in API-review.

[ChangeLog][QtBluetooth] QBluetoothUuid now provides all the
constructors that QUuid also provides (incl. from GUID and
QAnyStringView).

Change-Id: I2cb76a9a9cbc47f2fd4e637bdb41233d1ffc6972
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit aff0acdeee12acf6a85f479f160a0d4776f79d7c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>