QBluetoothUuid: magic_numbers -= 50
Prevent code readers from having to compare the long lists of magic
numbers in the various bluetooth-specific UUID constructors; they're
all the same as the ones used in the quint32 ctor, so use C++11 ctor
delegation to make all ctors call the quint32 one.

This calls into question the need of the quint16 ctor, seeing as the
quint32 ctor could just do the former's job with an implicit
conversion. But removing it would allow users to pass any integral
type to the QBluetoothUuid ctor, which is not the case atm, because
the two integral ctors make each other ambiguous unless the arguments
match either parameter types exactly. So leave it in.

Pick-to: 6.6
Change-Id: I40ddc47149cbb0526ad6488f32cc7f203e70a6b3
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>