Allow to set user information after detection started for non NDEF tags
Prior to the NDEF implementation one could change the user information
after the target detection started. The limitation seems to only effect
NDEF tags, so all other tags should still be able to set the
information at a later time. This allows to inform the user about the
current progress of the process or inform about errors.

Fixes: QTBUG-116563
Pick-to: 6.5 6.6
Change-Id: If82ea8c2056eb4c425626b06dac98462e3f0070c
Reviewed-by: Lars Schmertmann <lars.schmertmann@governikus.de>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>