Clear any pendingJob of BT LE controller in case of disconnect
When using QLowEnergyController in Client role and the remote
peripheral disconnects, we clear any current resources. However
it seems we didn't clear the pendingJob variable.

If the pendingJob variable remains non-null, then the next time
we use the controller, the "performNextIO()" function will return
early as it waits for the previous job to complete. However in this
case, as it is a remnant from a previous run, it never completes,
and the application's QLowEnergyService::discoverServiceDetails()
will silently do nothing (the new job is put into job queue, but
never attended, because of this pendingJob not completing).

Pick-to: 6.5
Fixes: QTBUG-115370
Change-Id: Idd14efcc1469c155ac46c6a35336e04fc5bd6aa9
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit cc337336fd51a13afc4ea09acf52c746fc747dbf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>