Add thread protection to Android static string cache
The string cache is accessed in the Qt Android main activity
thread, as wells as in JNI callbacks which execute in arbitrary
Java threads.

For clarity: the usage in code looks like this:
"valueForStaticField<QtJniTypes::XYZ"
There is a small chance that the cache is accessed concurrently.

Pick-to: 6.5
Change-Id: I34c26e22347f65614b18a000b5d61f0737e3b49e
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 62045a447c688935091ce6b9fb0cc8a0ce03343c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>