Android Bluetooth: Change static string cache to use QString
Relates to change 944200b5a9705a7617f82cdaf5caf8932380aba4 which
effectively terminates the application with a warning
if QJniObjects are accessed from different thread they were created in.
QJniObjects are not inherently thread-safe, as it is not
easily possible to synchronize the access with Java-side threads.

This change makes the internal string cache to use QStrings instead
of QJniObjects. The strings that are cached are constants as per
Android documentation, and don't change at runtime. The runtime
query and related caching of these strings is more of a measure
to protect against changes between eg. Android releases.

Change-Id: I76fa607ab6062da504071f3bb0161a30c27361e8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>