Silence expected java exceptions
Since we expect the calls to "connect" and "close" in
checkIsTargetLost() to fail with IOException when the target was lost,
we dont need to print the exceptions as log warning.

This restores the behavior like it was before the port to the new Qt 6
Android APIs (7d2c6ae66fd61be56b35ae0c59354b695a02d6c7).

Pick-to: 6.6 6.5
Change-Id: Ic9e14a58827f8b24865202012708644b903d5d70
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>