Revert "Remove support of Linux in Qt NFC"
This reverts commit 4282160a843a01457d71c995d870ab447ada655d.

Adapt to changes in QtNfc:
- omit .pri and .pro files from the qmake build system
- add re-added source files to CMake build system
- omit documentation about message handlers (no place for it)
- adjust documentation elsewhere to match current practices
- remove autogenerated DBus interfaces, generate them at compile
  time instead
- remove unused neard DBus interface XML files
- adjust code to current QtNfc API
- add the "we mean it" comment to just re-added private headers
- replace old license headers with equivalent SPDX license headers

If the PCSClite dependencies are found, the PCSClite backend needs
to be explicitly disabled to allow enabling the neard backend.
That means that the status quo regarding which backends are built
is largely maintained.
Compiling, not to mention runtime selection or simultaneous use,
of more than one backend continues to be unsupported, as it has
always been. It seems like neard is ~only used on embedded
systems, where a custom Qt build is par for the course. Debian
last shipped neard in Buster, currently oldoldstable. Ubuntu last
shipped it... at a time that is not covered anymore in its package
search. Most other large distributions do package neard, though.
I found the following that do: Suse, Fedora, Arch, Gentoo, Yocto.

Task-number: QTBUG-97943
Change-Id: I1b4541c49bb5855800fd949bdc20e200967e58b5
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>