NdefEditor: do not clear the existing message before reading
The read operation can be cancelled, or might fail for some reason.
In this case, the user does not expect the existing entries to be
deleted, so do not clear the message before reading.

If the read is successful, the existing message is replaced by the
new one, so no action is required here.

Update the example documentation to clarify this behavior.

Amends 115b4d947adf90a706f76de831f34960e3c62a9c.

Fixes: QTBUG-120410
Pick-to: 6.7 6.6 6.5
Change-Id: Ibaf00fb3c661abcc39b5d758e57ee5a837c0e270
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>