Unregister Bluetooth profile UUID after connection
The Bluetooth profile UUID is registered on DBus the whole time while
the device is connected, this prevent new connection from happening,
which also need to register to the same UUID on DBus.

The change is to unregister Bluetooth profile UUID after connection, as
it's no longer needed for the current connected device.

This will allow new devices that trying to connect to register the same
Bluetooth profile UUID On DBus.

[ChangeLog][QtBluetooth][Linux/BlueZ] Unregister Bluetooth profile UUID after connection,
so it can be registered for connecting to new devices.

Fixes: QTBUG-82413
Change-Id: I0e0b4cd303b32213725638efb323e6d00cb7b9e5
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>