iOS: Fix error signal
Currently targetLost is used before reportError. So the QNearFieldTarget
will be deleted by the user and the slot connected to the error signal
will never be called. In addition reportError use a QueuedConnection so
we need to do that for targetLost too and change the order.

Because targetLost is always used after invalidate
we can add emitting the signal to invalidate.

Pick-to: 6.7
Change-Id: I070f6bb705ccff0feacac7f58264ff6ea238bc89
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>