JNI: prepare for standard Java types getting declared in qtjnitypes.h
Only declare the standard types locally if the
QT_DECLARE_JNI_CLASS_STANDARD_TYPES preprocessor symbol is undefined.
That allows us to merge the relevant changes in each submodule
independently, without having to orchestrate the submodule update.

Once all changes are in and the submodule update is through, remove
the declarations entirely.

Pick-to: 6.8
Change-Id: I01a9a13ae7a1c56809d2755eb2318677264e33f0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>