Suppress bogus warning from MinGW
When trying to force asserts, MinGW starts complaining about calling
qFromUnaligned<uint16_t>() with a static QBA::_empty member:

  error: array subscript 'short unsigned int[0]' is partly outside
         array bounds of 'const char [1]' [-Werror=array-bounds=]

However, the code explicitly checks that QBA size is >= 2 before
going into this branch, so it looks like an incorrect disgnostics
from MinGW.

We already had the similar warnings in other places, and
qtbase/de005e7976e4b110d6b5f32a8bab5141713acb7b suggests that the
proper fix is to cache the QBA::size().

However, applying this fix alone does not solve the problem, so
explicitly create a QByteArrayView on the response, and use it
in all the operations, except for the final assignment.

Fixes: QTBUG-129763
Pick-to: 6.8 6.5
Change-Id: I1e7ce87801320a0960fded208a41984e7668dfb0
Reviewed-by: Marc Mutz <marc.mutz@qt.io>