BlueZ: use new qdbusxml2cpp namespace option
After qtbase/351b7a31aa24a704dba09121d91cb34190892315 implemented the
feature of wrapping generated classes into custom namespaces, use this
approach for all the generated proxies.

This commit supersedes e3a72c6017735448c4b074e905139c65dc8768a1,
because this is a better solution for bugs like QTBUG-133553.

The work on this patch also revealed that there is no script to
generate the B-Dus adapters. But that is a task for a separate patch.

As a drive-by: fix the DBusObjectManagerInterface xml file.

Pick the change to older branches, because we are pushing the generated
files directly, and they already require a bit of manual adjustments.
Just need to make sure that the cherry-picks wouldn't have the
unsupported option in the shell script that generates the files.

Changes for 6.9: Dropped the new --namespace option from the script
and added a comment that the namespace should be added manually.

Task-number: QTBUG-134305
Pick-to: 6.8
Change-Id: I9be9c184b1ec97893196b0acde74e0f7827ea9fc
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit a4ff2497ecee8aa6046ed0aa558f736dc9f614c5)