BlueZ: use custom namespace for D-Bus adaptors
Similarly to the previous commit that introduced this feature for
interfaces.

Also extend the bash script to generate the adaptors and rename
some functions in it.

Similarly to the previous commit, pick it to older branches
because we're pushing the generated files directly into the repo.
We need to make sure that the cherry-picked change does not
include the unsupported qdbusxml2cpp option in the shell script.

Task-number: QTBUG-134305
Pick-to: 6.9 6.8
Change-Id: I5176742ba28fd1fc6da03995ef5d3f5ea1d80dac
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>