BlueZ: use custom namespace for D-Bus adaptors
Similarly to the previous commit that introduced this feature for
interfaces.

Also extend the bash script to generate the adaptors and rename
some functions in it.

Similarly to the previous commit, pick it to older branches
because we're pushing the generated files directly into the repo.
We need to make sure that the cherry-picked change does not
include the unsupported qdbusxml2cpp option in the shell script.

Changes for 6.9: Dropped the new --namespace option from the script
and added a comment that the namespace should be added manually.

Task-number: QTBUG-134305
Change-Id: I5176742ba28fd1fc6da03995ef5d3f5ea1d80dac
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 1c9c9fce01240310531574d07d00cecec682cb38)
(cherry picked from commit 5b8e727986b54354d847a595a48dbd226cfecc86)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>