NdefEditor: fix B2Qt cross-builds on Windows
The example was generating a qml module named NdefEditor and a binary
named ndefeditor. During the build, the build tools create a
subdirectory named after the qml module in the build dir. As a result,
on case-insensitive filesystems (i.e. on Windows) this was leading to a
linker error, because the linker failed to create a binary (since there
already was a directory with the same name).
The issue was only observed during cross-compilation, because native
binaries get the ".exe" suffix.

Fix it by renaming the qml module to NdefEditorModule.

Amends 115b4d947adf90a706f76de831f34960e3c62a9c.

Fixes: QTBUG-133788
Task-number: QTQAINFRA-6985
Pick-to: 6.9 6.9.0 6.8 6.8.3 6.5
Change-Id: I1187474b07fe7c676612e79064f56d6711984c72
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>