Mark android/devicediscoverybroadcastreceiver.cpp as security-critical
The DeviceDiscoveryBroadcastReceiver::retrieveDeviceInfo() method
parses the incoming scan data.

The  data is received from the underlying Android OS APIs. However, I
do not consider it to be trusted, because we use the
ScanRecord.getBytes() Android API, which seems to return the raw data,
as received from the BT device.

Task-number: QTBUG-135467
Pick-to: 6.9 6.8
Change-Id: I01f176f4ab41da9dbf0b38793ad001c2f4b79168
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>