Mark bluez/hcimanager.cpp as security-critical
The handleHciEventPacket(), handleHciAclPacket(), and
handleLeMetaEvent() methods parse the data that is received as HCI
events.

The HCI and L2CAP protocols are implemented at the Linux kernel
level, so we can trust that the protocol-specific parts of the data
are valid.
However, we cannot be sure that the BT payload is anyhow verified,
so let's still play safe and consider these parsers security-critical.

Task-number: QTBUG-135467
Pick-to: 6.8
Change-Id: I3c7e67ca7f9314ee6883829dfd9ed05fd3d9967d
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 9ef2307c34f711234e98f26f3ab564e79fcc7ef3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>