HciManager::handleLeMetaEvent(): check size before accessing the data
The function was only taking a pointer to the data buffer and accesing
it without actually making sure that it has enough meaningful bytes.

Pass the actual buffer size as a second parameter to the function.
Print a warning and exit early if there is not enough bytes to parse
the data.

As in other commits, this cannot be considered a security issue,
because the provided buffer is pre-allocated on the stack, and it has
enough bytes to avoid reading past the end of the buffer. In worst
case we could get garbage data from the buffer.

Pick-to: 6.5
Change-Id: I67cd40e4644beac19ccf2a5fd531469ae951a49e
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 32ca71d6cca413c4521728b524d50d4e76572639)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit cada7d10b33d882a2ea9d0cee8b6bc6216a4d57f)