Mark qbluetoothservicediscoveryagent_bluez.cpp as security-critical
The QBluetoothServiceDiscoveryAgentPrivate::runExternalSdpScan()
method uses QProcess to execute an external binary (sdpscanner).
The binary itself is trusted, but we need to make sure that the path
to the binary is correct -> security-critical.

There are also methods in this file that parse the XML data received
from sdpscanner, but that data should be considered trusted (because
the sdpscanner tool is a part of Qt).

Task-number: QTBUG-135467
Pick-to: 6.9 6.8
Change-Id: Ia2ec436ad653c902e138813bb288c7a92e608d86
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>