Mark qleadvertiser_bluez.cpp as security-significant
The QLeAdvertiserBluez::handleCommandCompleted() method parses the
answers to the commands that we send. The answers are received from
the HciManager, which is considered to be security-critical.

However, the commands are only related to setting up the local
advertising. So, they are processed by the kernel part of the
Bluetooth stack and shouldn't have any data from remote devices.
As a result, we can trust that the data in the replies conforms
to the HCI protocol.

Task-number: QTBUG-135467
Pick-to: 6.9 6.8
Change-Id: Iad4978ef06286df102ccc0e563d4adf5d23b5bc6
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>