QLeAdvertiserBluez::handleCommandCompleted(): check data size
When reading the reply for the OcfLeReadTxPowerLevel command, check
that the data is not empty before accessing the first byte.

In practice the reply should always contain this byte, because this
command tries to read the power level from the local adapter, not from
some remote untrusted device.

Pick-to: 6.8 6.5
Change-Id: Iaa504f00a260b438e323e4e0f81ab84a31cd85b5
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 2343daa7943e241e227f223dd2df5322980b1a26)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>