QBluetoothSocketPrivateWinRT: remove unused variables
The m_pendingData variable was simply accumulating the incoming data,
but nothing was ever read from it.

The m_readMutex variable was supposed to guard access to m_pendingData.
Consequently, it was only used in one place when the data was added to
m_pendingData. It was also guarding the access to the rxBuffer, but
reading from the buffer is not protected by the mutex.

Both variables were added in 8a0d14f4933bcd933e1258731472479ec33ade6a
that introduced the Windows socket backend, and they were meaningless
even back then. It looks like a leftover from various design choices,
because the helper SocketWorker class has similar variables which are
actually used.

Pick-to: 6.9 6.8 6.5
Change-Id: I9fd5711f4e86929778b2a77bfa526b1390c8255a
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>