Android: Increase NFC transceive timeout
With Qt 6 a NFC transceive timeout of 2000ms was introduced with
the note, that "2000ms should be enough for most applications.".
See d15693d4ea2e06216f55facd504273c990b3814b for details. In our
application we noted that this is not enough for some cryptographic
operations especially when a password is checked. In case of a
wrong password some cards slow down the speed to make brute force
attacks more difficult. On July 2, 2024, we released a new version
for Android, where we increased the timeout to 3000ms with a Qt
patch. Since this we did not note any problems so it is time to
change this in the official Qt code too.

Change-Id: I0f0d977081cca01dcd9d0e54814fa92046107505
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 227718d5a7e5902d1935d7a05c56ef87eae3c360)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit ef3403112c744512c9081880d9d70bcd47a87388)